backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
flask-cors==4.0.0 